requirements.txt
requests==2.31.0
ffmpeg-python==0.2.0 